pyproject.toml
[project]
name = "ils-specs"
version = "0.1.0"
description = "ILS specification documents and validation tooling for CHPL"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    "mdformat>=0.7",
    "mdformat-frontmatter>=2.0",
    "mdformat-gfm>=0.3",
]